Added id check to removing 'like' and made adding 'like' return the new 'like'

+++ app/Http/Controllers/PostUserController.php
@@ -16,12 +16,18 @@ class PostUserController extends Controller
      */
     public function store(Request $request)
     {
+        // Get user and post
         $user = \Auth::user();
         $post = App\Post::find($request->post_id);
 
-        // Attach the post and user to eachother in the post_user table using attach()
+        // Attach the post and user to each other in the post_user table using attach()
         $user->likes()->attach($post);
 
+        // Get the user like
+        $like = $user->likes()->where(['post_user.user_id' => $user->id, 'post_user.post_id' => $request->post_id])->get();
+
+        // Return like
+        return $like;
     }
 
 
@@ -34,14 +40,21 @@ class PostUserController extends Controller
     public function destroy($id)
     {
 
+        // Get user
         $user = \Auth::user();
-        $like = $user->likes()->where(['post_user.user_id' => \Auth::user()->id, 'post_user.post_id' => $id])->get();
 
-//        $user->likes()->detach($like->id);
-//        $like->delete();
+        // Get user like
+        $like = $user->likes()->where(['post_user.user_id' => $user->id, 'post_user.post_id' => $id])->get();
 
-        \DB::delete('delete from post_user where user_id=? and post_id=?', array($user->id, $id));
+        // Check if the like is valid
+        if($like != null) {
 
-        return $like;
+            // Raw database command to delete the entry
+            \DB::delete('delete from post_user where user_id=? and post_id=?', array($user->id, $id));
+
+            return $like;
+        }
+
+        return response('Unauthorized', 403);
     }
 }
